% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand_ocean}
\alias{expand_ocean}
\title{A function to to add a row for each instance where a string occurs in a column}
\usage{
expand_ocean(data, column_name)
}
\arguments{
\item{data}{The data frame containing the rows to expand}

\item{column_name}{The column name holding the string to expand the data over in quotes (e.g. "Ocean").}
}
\description{
A function to to add a row for each instance where a comma-seperated string occurs in a column (e.g. if a cell contains the string "Atlantic, Pacific, Arctic
this suction will replace the row with three rows containing "Atlantic", "Pacific" and "Arctic"). This is useful when compiling summary
statistics for species that occur in multiple oceans.
}
\examples{
expand_ocean(data, "ocean")
}
\keyword{common}
\keyword{fish,}
\keyword{forage}
\keyword{life}
\keyword{literature}
\keyword{name,}
\keyword{species}
\keyword{stage,}
\keyword{summary}
