% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{how_many_times}
\alias{how_many_times}
\title{A function to determine how many times a pressure occurs in a column}
\usage{
how_many_times(pressures, parameter_columns)
}
\arguments{
\item{pressures}{The name of the pressure to be counted}

\item{parameter_columns}{The column(s) where the occurences of the pressure will be counted}
}
\description{
This function counts the number of instances of a pressure (word match) in a column where each row can contain
multiple pressures
}
\examples{
how_many_times("salinity",data[,22:35])
}
\keyword{drivers,}
\keyword{fish,}
\keyword{forage}
\keyword{literature}
\keyword{pressures,}
\keyword{summary}
