% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LHS_how_many_times}
\alias{LHS_how_many_times}
\title{A function to determine how many times a lifestage occurs in a column}
\usage{
LHS_how_many_times(lifestage, parameter_columns)
}
\arguments{
\item{lifestage}{The name of the lifestage to be counted}

\item{parameter_columns}{The column(s) where the occurences of the lifestage will be counted}
}
\description{
This function counts the number of instances of a lifestage (word match) in a column where each row can contain
multiple lifestages
}
\keyword{common}
\keyword{fish,}
\keyword{forage}
\keyword{life}
\keyword{literature}
\keyword{name,}
\keyword{species}
\keyword{stage,}
\keyword{summary}
