% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lifestage_table}
\alias{lifestage_table}
\title{A function to make a table of the number of times a life history stage occurs in the column "Life stage studied (Adult, Juvenile, Larval, Egg)"}
\usage{
lifestage_table(lifestage, parameter_columns)
}
\arguments{
\item{lifestage}{The name of the life stage to be counted}

\item{parameter_column}{The column where the occurences of the the life history stage will be counted}
}
\description{
This function to make a table of the number of times a life history stage occurs in that column.
}
\keyword{common}
\keyword{fish,}
\keyword{forage}
\keyword{life}
\keyword{literature}
\keyword{name,}
\keyword{species}
\keyword{stage,}
\keyword{summary}
