% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pressure_table}
\alias{pressure_table}
\title{A function to make a table of the number of times a pressure occurs in a set of columns}
\usage{
pressure_table(pressures, parameter_columns)
}
\arguments{
\item{pressures}{The name of the pressure to be counted}

\item{parameter_columns}{The column(s) where the occurences of the pressure will be counted}
}
\description{
This function to make a table of the number of times a pressure occurs in a set of columns.
}
\examples{
pressure_table(c("saliity","temperature"),data[,22:32])
}
\keyword{drivers,}
\keyword{fish,}
\keyword{forage}
\keyword{literature}
\keyword{pressures,}
\keyword{summary}
